+++ src/xtd.core.native.unix/src/xtd/native/unix/speech_synthesizer.cpp
@@ -18,7 +18,7 @@ using namespace xtd::native;
 namespace {
   struct speech_synthesizer_data {
     string say_cmd_file_name;
-    intptr_t process_handle = 0;
+    intmax_t process_handle = 0;
   };
   
   static string get_temp_path() {
@@ -34,36 +34,36 @@ namespace {
   }
 }
 
-intptr_t speech_synthesizer::create() {
+intmax_t speech_synthesizer::create() {
   speech_synthesizer_data* data = new speech_synthesizer_data {get_unique_speak_cmd_file_name(), 0};
   ofstream cmd_file;
   cmd_file.open(data->say_cmd_file_name);
   cmd_file << "spd-say \"$*\"\n";
   cmd_file.close();
   native::file_system::set_permissions(data->say_cmd_file_name, FILE_PERMISSIONS_OWNER_ALL);
-  return reinterpret_cast<intptr_t>(data);
+  return reinterpret_cast<intmax_t>(data);
 }
 
-void speech_synthesizer::destroy(intptr_t handle) {
+void speech_synthesizer::destroy(intmax_t handle) {
   unlink(reinterpret_cast<speech_synthesizer_data*>(handle)->say_cmd_file_name.c_str());
   delete reinterpret_cast<speech_synthesizer_data*>(handle);
 }
 
-void speech_synthesizer::pause(intptr_t handle) {
+void speech_synthesizer::pause(intmax_t handle) {
 
 }
 
-void speech_synthesizer::resume(intptr_t handle) {
+void speech_synthesizer::resume(intmax_t handle) {
 
 }
 
-void speech_synthesizer::speak(intptr_t handle, const string& text_to_speak) {
+void speech_synthesizer::speak(intmax_t handle, const string& text_to_speak) {
   speak_async(handle, text_to_speak, [] {});
   int32_t exit_code = 0;
   native::process::wait(reinterpret_cast<speech_synthesizer_data*>(handle)->process_handle, exit_code);
 }
 
-void speech_synthesizer::speak_async(intptr_t handle, const string& text_to_speak, std::function<void()> on_speak_completed) {
+void speech_synthesizer::speak_async(intmax_t handle, const string& text_to_speak, std::function<void()> on_speak_completed) {
   reinterpret_cast<speech_synthesizer_data*>(handle)->process_handle = native::process::shell_execute("", reinterpret_cast<speech_synthesizer_data*>(handle)->say_cmd_file_name, text_to_speak, "", PROCESS_WINDOW_STYLE_HIDDEN);
   thread wait_process_thread([on_speak_completed, handle] {
     int32_t exit_code = 0;
@@ -73,6 +73,6 @@ void speech_synthesizer::speak_async(intptr_t handle, const string& text_to_spea
   wait_process_thread.detach();
 }
 
-void speech_synthesizer::stop(intptr_t handle) {
+void speech_synthesizer::stop(intmax_t handle) {
 
 }
